+++ packages/client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- CLI: remove the descriptors folder if it exists before attempting to create it.
+
 ## 0.1.1 - 2024-04-11
 
 ### Fixed